+++ src/Tienda/Productos/TipoProductos/Bebida.java
@@ -14,7 +14,7 @@ public class Bebida extends Producto{
             String identificador,
             String descripcion,
             int stock,
-            double precioUnitario,
+            float precioUnitario,
             float porcentajeGanancia,
             boolean disponible,
             float gradoAlcohol,
