+++ src/Tienda/Productos/TipoProductos/ProductoLimpieza.java
@@ -9,7 +9,7 @@ public class ProductoLimpieza extends Producto {
             String identificador,
             String descripcion,
             int stock,
-            double precioUnitario,
+            float precioUnitario,
             float porcentajeGanancia,
             boolean disponible,
             TipoAplicacionLimpieza tipoDeAplicacion)
